energy_news_project/requirements.txt
feedparser
requests
beautifulsoup4
transformers
torch
deep-translator
python-telegram-bot

